refactor(tau-ops): split channel_store_admin helper modules (#1257)
